pyproject.toml
[tool.poetry]
name = "ankimate"
version = "0.1.0"
description = ""
authors = ["adaxiang <adaxiang3104@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
english-ipa = "0.2.1"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"
lxml = "^5.2.2"
aiohttp = "^3.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
